+++ src/Symfony/Component/Yaml/CHANGELOG.md
@@ -6,3 +6,5 @@ CHANGELOG
 
  * Yaml::parse() does not evaluate loaded files as PHP files by default
    anymore (call Yaml::enablePhpParsing() to get back the old behavior)
+ * Inline::parse() now permits the last parsed index to be determined
+   through an optional by-reference parameter